better storage naming; use GM_ storage if possible
I decided that using the namespace as a name was a bad idea if you ever write another script, so I now use the initials from the script name. And since I was already updating anyways, I thought to reverse my previous decision and continue to use GM_ storage if possible, so settings won't be lost unless necessary. 

I also added a shim for GM_info, which I neglected before. It currently exists in GM 4, but it may be removed in the future, hence the weird testing. 

Not sure how many edits I should do in one commit, but this one is tested (as much as the rest).